pyproject.toml
[project]
name = "knowledge-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]

[dependency-groups]
dev = [
    "asyncio>=4.0.0",
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
]
